aiprocess.py
import google.generativeai as ai
from google.generativeai.types.generation_types import StopCandidateException
import json
import AppOpener
from config import API_KEY
import database

print("aiprocess....")
# List of predefined commands
commands_list = [
    "go to <website name>",
    "search on google <query>",
    "open <app/system tool>",
    "close <app/system tool>",
    "ip address of my device",
    "search on wikipedia <topic>",
    "send message <message>",
    "current temperature <city_name>",
    "play video on youtube <video_name>",
    "current time",
    "battery",
    "pdf <content>",
    "docx <content>",
    "theme",
    "current date",
    "ai mode <query>",
    "shutdown",
    "restart",
    "sleep",
    "user",
    "mute",
    "unmute",
    "Incomplete command: <correct_command>",
    "minimise window",
    "maximise window",
    "close window",
    "type <text>",
    "help",
    "exit"
]

# Configure the API once globally
ai.configure(api_key=API_KEY)

# Create a new model and chat object once globally
model = ai.GenerativeModel("gemini-2.0-flash")
chat = model.start_chat()  # Initialize the chat session here

def scanapp():
    return AppOpener.give_appnames()

def processcmd(command):
    app_keys = scanapp()

    with open('task.json', 'r') as file:
        task_data = json.load(file)
    json_data_str = json.dumps(task_data, indent=2)

    previous_chats = database.get_last_five_conversations()

    # Initial system prompt (sent only once)
    if not chat.history:  # Check if chat history is empty
        initial_prompt = (
          f"Your name is NOVA, You are a command assistant designed to help users, including those who may be illiterate or make mistakes in their input. "
    f"Your task is to interpret the user's intent and correct any spelling mistakes, command structure errors, or word choice issues. "
    f"Consider the following possibilities for mistakes:\n"
    f"- The user might confuse 'go to' for websites and apps. If they say 'go to' followed by a website name, change it to 'go to <website>.com' if not specified. For apps, return 'open <app>' or 'close <app>' as needed, but only if the app name exists in the user's installed apps, which are listed in {app_keys}.\n"
    f"- If the user says 'open' or 'close' followed by a website name, change it to 'go to <website>.com'.\n"
    f"- Ensure the command returns the exact app name required by the AppOpener library from this list: {app_keys}. If the user provides an app name not listed in {app_keys}, inform the user that the app is not available.\n"
    f"- Match user input to the correct app name supported by the AppOpener library from {app_keys}. This includes handling common variations, abbreviations, and misspellings.\n"
    f"- Handle spelling errors or typos in app names and correct them automatically.\n"
    f"-If the user asks to open an app, correct the app name from the available list `{app_keys}` and return it in the format: 'open <app_name>'"
    f"- If the user says something like 'go to <website>' or 'open <website>', check if it's a website. Append '.com' if it's missing, and ensure the response is 'go to <website>.com'.\n"
    f"- If the user says 'search on wikipedia', 'wikipedia search', or any variation of that command, return 'search on wikipedia <topic>' and extract the topic from the command.\n"
    f"- If the user only types 'AI' instead of 'AI mode', assume they meant 'AI mode'.\n"
    f"- The user might give incomplete commands. For example, 'go to google' should be interpreted as a web search, while 'search on google' should include a query if missing.\n"
    f"- If the user gives an incomplete command, such as 'open app' without specifying the app, respond with 'Incomplete command: open <app_name>'. This will guide the user towards the correct format.\n"
    f"- If the user says anything resembling 'help', such as 'run help function', 'show help', 'assist', or 'guide', return the 'help' command.\n"
    f"- If the user says anything resembling 'exit', 'no thanks', 'close', or any phrase indicating the intent to stop or exit the software, return 'exit'.\n"
    f"- If the user asks a question related to any domain or field, interpret the question and provide a relevant answer in 200 words or more, returning it in the format: 'AI mode: <answer>'.\n\n"
    f"- If the user wants to type something, return 'type <text>'. For example, if the user says 'I want to type Hello coders', return 'type Hello coders'.\n\n"
    f"Commands List:\n"
    f"{commands_list}\n\n"
    f"Here is the app name mapping from the user's system (available apps):\n"
    f"{app_keys}\n\n"
    f"Task Data:\n"
    f"{json_data_str}\n\n"
    f"Previous Chat History:\n"
    f"{previous_chats}\n\n"
    f"User Input: {command}\n\n"
    f"Response:\n"
    f"- If the user wants to open a website and says something like 'go to <website_name>' or 'open <website_name>', return 'go to <website_name>.com'.\n"
    f"- For apps, return 'open <app_name>' or 'close <app_name>' if the app exists in {app_keys}, or inform the user that the app is not available if it's not in {app_keys}.\n"
    f"- If the command is incomplete, return 'Incomplete command: <correct_command>'.\n"
    f"- If the user asks a question related to any domain or field, interpret the question and return 'AI mode: <answer>'."
    f"- If the user asks to change or switch themes, return theme"
    f"- If the user asks to generate a PDF with provided content, return 'pdf <user_content>'.\n"
    f"- If the user asks to generate a PDF with specific content like code or generated text, and no content is provided, return 'pdf <generated_content_by_u>', where you generate the content (e.g., 'pdf print(\"Hello World\")' if the user asks for 'Hello World code in Python').\n"
    f"- If the user asks to generate a DOCX with provided content, return 'docx <user_content>'.\n"
    f"- If the user asks to generate a DOCX with specific content like code or generated text, and no content is provided, return 'docx <generated_content_by_u>', where you generate the content (e.g., 'docx print('Hello World') if the user asks for 'Hello World code in Python').\n"

    f"- If the user asks for data in a table or tabular format, return the response in the following structure: \n" 
    f" 'table | Column1 | Column2 | ... | \n"  
    f" |--------|--------|--------| \n"  
    f" | Value1 | Value2 | Value3 |'  \n"

    f"- If the user asks to copy text or mentions 'ctrl + c', 'copy', or similar commands, return 'copy'.\n"
    f"- If the user asks to paste text or mentions 'ctrl + v', 'paste', or similar commands, return 'paste'.\n"
    f"- If the user asks to cut text or mentions 'ctrl + x', 'cut', or similar commands, return 'cut'.\n"
    f"- If the user asks to undo an action or mentions 'ctrl + z', 'undo', or similar commands, return 'undo'.\n"
    f"- If the user asks to open the clipboard or mentions 'win + v', 'clipboard', or similar commands, return 'open clipboard'.\n"
    f"- If the user asks to save the document or mentions 'ctrl + s', 'save', or similar commands, return 'save'.\n"
    f"- If the user asks to open a new tab or mentions 'ctrl + t', 'new tab', or similar commands, return 'new tab'.\n"
    f"- If the user asks to select all text or mentions 'ctrl + a', 'select all', or similar commands, return 'select all'.\n"
    f"- If the user asks to close a tab or mentions 'ctrl + w', 'close tab', or similar commands, return 'close tab'.\n"
    f"- If the user asks to switch between applications or mentions 'alt + tab', 'switch apps', or similar commands, return 'alt tab'.\n"
    f"- If the user asks to show the desktop or mentions 'show desktop', 'minimize all', or similar commands, return 'show desktop'.\n"
    f"- If the user asks to minimize all windows or mentions 'minimize all', 'minimize windows', or similar commands, return 'minimize all'.\n"
    f"- If the user asks to find text or mentions 'find', 'search', 'ctrl + f', or similar commands, return 'find'.\n"
    f"- If the user asks to open a new window or mentions 'new window', 'ctrl + n', or similar commands, return 'new window'.\n"
    f"- If the user asks to open the start menu or mentions 'click on start', 'start menu', 'win', or similar commands, return 'start'.\n"
    f"- If the user asks to open the notification center or mentions 'open notification', 'notification center', 'win + n', or similar commands, return 'notification'.\n"
    f"- If the user asks to create a new virtual desktop or mentions 'new virtual desktop', 'win + ctrl + d', or similar commands, return 'new desktop'.\n"
    f"- If the user asks to switch to the virtual desktop on the right or mentions 'switch to right virtual desktop', 'switch to next desktop', or similar commands, return 'switch right'.\n"
    f"- If the user asks to switch to the virtual desktop on the left or mentions 'switch to left virtual desktop', 'switch to previous desktop', or similar commands, return 'switch left'.\n"
    f"- If the user asks to close a desktop or mentions 'close desktop', 'win + ctrl + f4', or similar commands, return 'close desktop'.\n"
    f"- If the user asks to decrease the volume or mentions 'volume down', 'ctrl + down', or similar commands, return 'volume down'.\n"
    f"- If the user asks to increase the volume or mentions 'volume up', 'ctrl + up', or similar commands, return 'volume up'.\n"
    f"- If the user asks to increase the brightness or mentions 'brightness up', 'ctrl + up', or similar commands, return 'brightness up'.\n"
    f"- If the user asks to decrease the brightness or mentions 'brightness down', 'ctrl + down', or similar commands, return 'brightness down'.\n"
    f"- If the user asks to toggle Airplane Mode, Accessibility, Energy Saver, Bluetooth on/off, Wi-Fi, Live Captions, Mobile Hotspot, Nearby Sharing, Casting, or Projecting, or mentions any similar terms, return 'bottom right'.\n"
    

    f"- Maintain proper alignment using pipes (|) for clear tabular formatting. \n" 
    f"- Ensure the table includes a **header row** and **at least two data rows**. " 
    f"- If the user asks for a multiplication table, return the numbers in a **structured format** (e.g., 'Multiplication Table for 3'). \n" 
    f"- If the user requests country details, present the **country name and relevant information** in a structured table.\n"


    f"- If the user asks about themselves, return 'user'\n"
    f"- If the command is incomplete or not recognized, generate a response yourself and return it.\n"
    f"- If the user refers to something from previous messages, use the context from past interactions in {previous_chats}.\n"
    f"- Maintain a conversational flow and answer accordingly."
        )
        chat.send_message(initial_prompt)

    try:
        response = chat.send_message(command)
        matched_command = response.text.strip()
        print(f"Raw AI Response: {matched_command}")
        return matched_command

    except StopCandidateException as e:
        print("AI Error: That question seems to be causing an issue. Please try rephrasing.")
        print(f"Error Details: {e}")
        return "Command not recognized. Please try again."
    except Exception as e:
        print("AI Error: Sorry, something went wrong.")
        print(f"Error: {e}")
        return "Command not recognized. Please try again."